Implement the main crate types
This commit adds most of the crates types and implementations, i.e. the
doc-examples of the top-level crate documentation all compile and run
fine. Thus the basic implementation is done.
The implementation is checked with a number of tests.

It is not complete, both the implementation and the documentation can be
improved.